# api/openapi.yaml
openapi: 3.1.0
info:
  title: MaxPosty API
  version: 1.0.0
  description: |
    Canonical HTTP contract shared by the MaxPosty backend and frontend.
    Browser sessions are carried by the secure `maxstudio_session` cookie.
servers:
  - url: https://api.maxposty.ru/api/v1
    description: Production
  - url: http://localhost:8080/api/v1
    description: Local development
...
paths:
...
  /posts:
    get:
      tags: [Posts]
      operationId: listPosts
      summary: List posts in the current workspace
      security:
        - sessionCookie: []
      parameters:
        - name: channel_id
          in: query
          schema:
            type: integer
            format: int64
            minimum: 1
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/PostStatus'
      responses:
        '200':
          description: Posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
        default:
          $ref: '#/components/responses/Problem'
    post:
      tags: [Posts]
      operationId: createPost
      summary: Create a draft
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditablePost'
      responses:
        '201':
          description: Created post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        default:
          $ref: '#/components/responses/Problem'
...
  /posts/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Posts]
      operationId: getPost
      summary: Get one post
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        default:
          $ref: '#/components/responses/Problem'
    put:
      tags: [Posts]
      operationId: replacePost
      summary: Save editable post fields
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostUpdate'
      responses:
        '200':
          description: Saved post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        default:
          $ref: '#/components/responses/Problem'
    patch:
      tags: [Posts]
      operationId: updatePost
      summary: Save selected editable post fields
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostUpdate'
      responses:
        '200':
          description: Saved post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      tags: [Posts]
      operationId: deletePost
      summary: Delete a local post
      security:
        - sessionCookie: []
      responses:
        '204':
          description: Post deleted
        default:
          $ref: '#/components/responses/Problem'
  /posts/{id}/duplicate:
    post:
      tags: [Posts]
      operationId: duplicatePost
      parameters:
        - $ref: '#/components/parameters/Id'
      security:
        - sessionCookie: []
      responses:
        '201':
          $ref: '#/components/responses/Post'
        default:
          $ref: '#/components/responses/Problem'
  /posts/{id}/generate-image:
    post:
      tags: [AI]
      operationId: generatePostImage
      summary: Generate and attach a post image
      parameters:
        - $ref: '#/components/parameters/Id'
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeneratePostImageInput'
      responses:
        '200':
          $ref: '#/components/responses/Post'
        default:
          $ref: '#/components/responses/Problem'
...
  /posts/{id}/publish:
    post:
      tags: [Posts]
      operationId: publishPost
      summary: Publish a post immediately
      parameters:
        - $ref: '#/components/parameters/Id'
      security:
        - sessionCookie: []
      responses:
        '200':
          $ref: '#/components/responses/Post'
        default:
          $ref: '#/components/responses/Problem'
  /posts/{id}/schedule:
    post:
      tags: [Posts]
      operationId: schedulePost
      summary: Schedule a post for a future time
      parameters:
        - $ref: '#/components/parameters/Id'
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [scheduled_at]
              properties:
                scheduled_at:
                  type: string
                  format: date-time
      responses:
        '200':
          $ref: '#/components/responses/Post'
        default:
          $ref: '#/components/responses/Problem'
...
  /posts/{id}/update-published:
    post:
      tags: [Posts]
      operationId: updatePublishedPost
      parameters:
        - $ref: '#/components/parameters/Id'
      security:
        - sessionCookie: []
      responses:
        '200':
          $ref: '#/components/responses/Post'
        default:
          $ref: '#/components/responses/Problem'
...
  /posts/{id}/publication:
    delete:
      tags: [Posts]
      operationId: deletePublication
      parameters:
        - $ref: '#/components/parameters/Id'
      security:
        - sessionCookie: []
      responses:
        '200':
          $ref: '#/components/responses/Post'
        default:
          $ref: '#/components/responses/Problem'
...
  /images/generate:
    post:
      tags: [AI]
      operationId: generateImage
      summary: Generate an image in the personal workspace
      description: Legacy personal-workspace alias. New clients should use the workspace-scoped endpoint.
      deprecated: true
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateImageInput'
      responses:
        '201': {$ref: '#/components/responses/MediaFile'}
        default: {$ref: '#/components/responses/Problem'}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: maxstudio_session
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
  responses:
    Problem:
      description: Request failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemEnvelope'
    Post:
      description: Updated post
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Post'
...
    MediaFile:
      description: Private workspace media metadata
      content: {application/json: {schema: {$ref: '#/components/schemas/MediaFile'}}}
...
  schemas:
    ProblemEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/Problem'
    Problem:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details: {}
...
    MediaFile:
      type: object
      required: [type, url, filename, mime_type, size, width, height]
      properties:
        type: {type: string, enum: [image, video]}
        url: {type: string}
        filename: {type: string}
        mime_type: {type: string}
        size: {type: integer, format: int64, minimum: 0}
        width: {type: integer, minimum: 0}
        height: {type: integer, minimum: 0}
        duration_ms: {type: integer, format: int64, minimum: 0}
...
    PostReviewStatus:
      type: string
      enum: [draft, in_review, changes_requested, approved]
...
    PostFormat:
      type: string
      enum: [markdown, html]
    PostStatus:
      type: string
      enum: [draft, scheduled, publishing, published, failed]
    PostLinkButton:
      type: object
      required: [text, url]
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 64
        url:
          type: string
          format: uri
    PostAttachment:
      type: object
      required:
        - id
        - type
        - position
        - url
        - processing_status
        - size_bytes
        - mime_type
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
        type:
          type: string
          enum: [image, video]
        position:
          type: integer
          minimum: 0
          maximum: 11
        url:
          type: string
        processing_status:
          type: string
          enum: [uploading, processing, ready, failed]
        size_bytes:
          type: integer
          format: int64
          minimum: 0
        mime_type:
          type: string
        width:
          type: [integer, 'null']
          minimum: 1
        height:
          type: [integer, 'null']
          minimum: 1
        duration_ms:
          type: [integer, 'null']
          format: int64
          minimum: 0
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    EditablePost:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
        content:
          type: string
          maxLength: 4000
        format:
          $ref: '#/components/schemas/PostFormat'
        channel_id:
          type: [integer, 'null']
          format: int64
        image_url:
          type: [string, 'null']
        image_prompt:
          type: [string, 'null']
        link_buttons:
          type: array
          maxItems: 3
          items:
            $ref: '#/components/schemas/PostLinkButton'
        notify:
          type: boolean
        disable_link_preview:
          type: boolean
        scheduled_at:
          type: [string, 'null']
          format: date-time
    PostUpdate:
      allOf:
        - $ref: '#/components/schemas/EditablePost'
        - type: object
          properties:
            expected_updated_at:
              type: string
              format: date-time
    Post:
      type: object
      required:
        - id
        - title
        - content
        - format
        - status
        - attachments
        - link_buttons
        - notify
        - disable_link_preview
        - max_is_pinned
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
        workspace_id:
          type: string
        title:
          type: string
        content:
          type: string
        format:
          $ref: '#/components/schemas/PostFormat'
        status:
          $ref: '#/components/schemas/PostStatus'
        channel_id:
          type: [integer, 'null']
          format: int64
        image_url:
          type: [string, 'null']
        image_prompt:
          type: [string, 'null']
        attachments:
          type: array
          maxItems: 12
          items:
            $ref: '#/components/schemas/PostAttachment'
        link_buttons:
          type: array
          items:
            $ref: '#/components/schemas/PostLinkButton'
        notify:
          type: boolean
        disable_link_preview:
          type: boolean
        scheduled_at:
          type: [string, 'null']
          format: date-time
        max_message_id:
          type: [string, 'null']
        max_message_url:
          type: [string, 'null']
          format: uri
        max_views:
          type: [integer, 'null']
          format: int64
          minimum: 0
        max_stats_synced_at:
          type: [string, 'null']
          format: date-time
        max_is_pinned:
          type: boolean
        last_error:
          type: [string, 'null']
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        published_at:
          type: [string, 'null']
          format: date-time
        review_status:
          $ref: '#/components/schemas/PostReviewStatus'
        current_revision_id:
          type: [integer, 'null']
          format: int64
    GenerateImageInput:
      type: object
      required: [prompt]
      properties:
        prompt:
          type: string
          minLength: 1
          maxLength: 2000
        size:
          type: string
          enum: [1024x1024, 1536x1024, 1024x1536]
          default: 1024x1024
        quality:
          type: string
          enum: [low, medium, high]
          default: medium
    GeneratePostImageInput:
      type: object
      properties:
        prompt:
          type: string
          maxLength: 2000
          description: Optional override; a blank or omitted value uses the post image prompt.
        size:
          type: string
          enum: [1024x1024, 1536x1024, 1024x1536]
          default: 1024x1024
        quality:
          type: string
          enum: [low, medium, high]
          default: medium
...